+++ internal/pkg/config/ssh.go
@@ -1,5 +1,5 @@
 package config
 
 type SSHConf struct {
-	KeyTypes []string `yaml:"key types,omitempty" default:"[\"rsa\",\"dsa\",\"ecdsa\",\"ed25519\"]"`
+	KeyTypes []string `yaml:"key types,omitempty" default:"[\"ed25519\",\"ecdsa\",\"rsa\",\"dsa\"]"`
 }